ui: phase 5k — extract Normal mode handler
Move handleNormalMode (~200 lines, the largest mode by far) to
mode_normal.go as a free function. Swap the modeHandlers map
entry and update dispatchModeKey's fallback to use the free
function. Test shim added.

Update the package-private compile-time assertion at the bottom
of mode_handlers.go to reference the free-function form
(`var _ modeHandler = handleNormalMode`) instead of the now-
gone method value.

24/24 ui packages green, vet clean.
